Time: 100 ms (6.54%), Space: 44.1 MB (17.8%) - LeetHub

+++ 2279-maximum-split-of-positive-even-integers/2279-maximum-split-of-positive-even-integers.cpp
@@ -3,7 +3,7 @@ public:
     vector<long long> maximumEvenSplit(long long finalSum) {
         if (finalSum%2==1) return {};
         vector<long long> ans;
-        long long i=2;
+        int i=2;
         while (2*i+2<=finalSum)
         {
             ans.push_back(i);
